requirements.txt

#Django
Django==2.2


# Rest apis
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
pytz==2018.5
requests==2.21.0
serpy==0.3.1

# Python-PostgreSQL Database Adapter
psycopg2==2.8.2

# Time zones support
pytz==2018.5

django-cors-middleware==1.3.1

# Our Web server (light & fast)
gunicorn==19.9.0

django-environ==0.4.1
django-heroku==0.3.1
ipython==7.5.0

# heroku needs it
django_extensions
django-storages
dj-database-url==0.5.0
whitenoise==4.1.2